requirements.txt
folium
pand
streamlit-folium
streamlit
plotly
pand
psycopg2
